fix removal cycle (again)

+++ backup.sh
@@ -16,15 +16,22 @@ gsutil cp $FILENAME_PATH gs://$BACKUP_BUCKET/$DATE.sql
 echo "Dump uploaded"
 
 gsutil ls gs://$BACKUP_BUCKET/ | while read -r file; do
+    file=$(echo "$file" | tr -d '\r\n')
     filename=$(basename "$file" .sql)
     if [[ $filename =~ ^[0-9]{2}-[0-9]{2}-[0-9]{4}$ ]]; then
-        file_timestamp=$(date -d "$(echo $filename | sed 's/\([0-9]\{2\}\)-\([0-9]\{2\}\)-\([0-9]\{4\}\)/\2\/\1\/\3/')" +%s)
-        current_timestamp=$(date +%s)
-        age_days=$(( (current_timestamp - file_timestamp) / 86400 ))
+        file_date=$(date -d "$(echo $filename | sed 's/\([0-9]\{2\}\)-\([0-9]\{2\}\)-\([0-9]\{4\}\)/\2\/\1\/\3/')" +%s 2>/dev/null)
+        if [ $? -ne 0 ]; then
+            echo "Error parsing date from filename: $filename"
+            continue
+        fi
+        current_date=$(date +%s)
+        age_days=$(( (current_date - file_date) / 86400 ))
         if [ $age_days -gt $BACKUP_RETENTION_DAYS ]; then
             gsutil rm "$file"
             echo "Deleted $file, $age_days days old"
         fi
+    else
+        echo "Skipping invalid filename format: $filename"
     fi
 done
